+++ contracts/packages/@rarible/royalties/contracts/IRoyaltiesProvider.sol
@@ -6,5 +6,7 @@ pragma abicoder v2;
 import "./LibPart.sol";
 
 interface IRoyaltiesProvider {
-    function getRoyalties(address token, uint tokenId) external returns (LibPart.Part[] memory);
+    function getRoyalties(address token, uint256 tokenId)
+        external
+        returns (LibPart.Part[] memory);
 }
