+++ contracts/packages/@rarible/royalties/contracts/RoyaltiesV1.sol
@@ -3,8 +3,16 @@
 pragma solidity >=0.6.2;
 
 interface RoyaltiesV1 {
-    event SecondarySaleFees(uint256 tokenId, address[] recipients, uint[] bps);
+    event SecondarySaleFees(
+        uint256 tokenId,
+        address[] recipients,
+        uint256[] bps
+    );
 
-    function getFeeRecipients(uint256 id) external view returns (address payable[] memory);
-    function getFeeBps(uint256 id) external view returns (uint[] memory);
+    function getFeeRecipients(uint256 id)
+        external
+        view
+        returns (address payable[] memory);
+
+    function getFeeBps(uint256 id) external view returns (uint256[] memory);
 }
